combined turretTarget files to be toggleable
