+++ src/main/java/frc5124/robot/commands/turretcommands/toggleLimeLight.java
@@ -17,8 +17,7 @@ public class toggleLimeLight extends CommandBase {
   /**
    * Creates a new toggleLimeLight.
    */
-  public toggleLimeLight() {
-  }
+  public toggleLimeLight() {}
 
   // Called when the command is initially scheduled.
   @Override
@@ -29,8 +28,7 @@ public class toggleLimeLight extends CommandBase {
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {
-  }
+  public void execute() {}
 
   // Called once the command ends or is interrupted.
   @Override
